Standardized more code for compilation on VC8, cleaned code (wrong indentation, etc.) and some bug fixes (untested like the original code there)
Standardized more code: 
modified   ctp2_code/ai/CityManagement/governor.cpp
modified   ctp2_code/ai/diplomacy/diplomat.cpp

Standardized and cleaned up code (replaced counter variable names by more readable standart names):
modified   ctp2_code/gs/gameobj/ArmyData.cpp

Cleaned up code (fixed wrong indentation, added missing header comments, made comments more readable, counter variables)
modified   ctp2_code/ai/diplomacy/regardevent.cpp
modified   ctp2_code/gs/gameobj/bldque.cpp
modified   ctp2_code/gs/gameobj/buildingutil.cpp
modified   ctp2_code/gs/gameobj/buildingutil.h
modified   ctp2_code/gs/gameobj/CityEvent.cpp
modified   ctp2_code/gs/gameobj/Player.cpp
modified   ctp2_code/gs/gameobj/UnitData.cpp
modified   ctp2_code/gs/gameobj/wonderutil.cpp
modified   ctp2_code/gs/gameobj/wonderutil.h
modified   ctp2_code/gs/slic/SlicFrame.cpp
modified   ctp2_code/gs/slic/SlicStruct.h
modified   ctp2_code/ui/aui_common/aui_blitter.cpp
modified   ctp2_code/ui/aui_common/aui_listbox.cpp
modified   ctp2_code/ui/aui_common/aui_screen.cpp
modified   ctp2_code/ui/aui_ctp2/ctp2_TabGroup.cpp
modified   ctp2_code/ui/aui_ctp2/SelItemClick.cpp
modified   ctp2_code/ui/aui_utils/primitives.cpp
modified   ctp2_code/ui/interface/armymanagerwindow.cpp
modified   ctp2_code/ui/interface/creditsscreen.cpp
modified   ctp2_code/ui/interface/loadsavescreen.cpp
modified   ctp2_code/ui/interface/sciencevictorydialog.cpp
modified   ctp2_code/ui/netshell/allinonewindow.cpp

Fixed code, cleaned up code (wrong indentiation, counter variables, line wrapping in comments)
modified   ctp2_code/gs/gameobj/CityData.cpp

Removed unnecessary default values for bits, removed the bit marker from flags that are not bits according to the source code and assigned to them a default value if necessary. Fixed indentation, made comments more readable:
modified   ctp2_code/gs/newdb/building.cdb
modified   ctp2_code/gs/newdb/citystyle.cdb
modified   ctp2_code/gs/newdb/Difficulty.cdb
modified   ctp2_code/gs/newdb/government.cdb
modified   ctp2_code/gs/newdb/terrain.cdb
modified   ctp2_code/gs/newdb/terrimprove.cdb
modified   ctp2_code/gs/newdb/unit.cdb

Standardized code and reverted to pre revision 575 state. This was not intended to be submitted but somehow it sipt through:
modified   ctp2_code/ui/interface/spnewgamediffscreen.cpp

git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@578 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gs/slic/SlicFrame.cpp
@@ -38,7 +38,8 @@
 // - Repaired crash with invalid input.
 // - Initialized local variables. (Sep 9th 2005 Martin Gühmann)
 // - Added database array access. (Sep 16th 2005 Martin Gühmann)
-// - Repaired crashes with game saved with original Activision executable. 
+// - Repaired crashes with game saved with original Activision executable.
+// - Standardized code. (May 29th 2006 Martin Gühmann) 
 //
 //----------------------------------------------------------------------------
 
@@ -557,7 +558,7 @@ BOOL SlicFrame::DoInstruction(SOP op)
 			Assert(sp >= 0);
 			sp = m_stack->Pop(type2, sval2);
 			Assert(sp >= 0);
-			sval3.m_int = (int)pow(Eval(type2, sval2), Eval(type1, sval1));
+			sval3.m_int = (int)pow(static_cast<double>(Eval(type2, sval2)), Eval(type1, sval1));
 			m_stack->Push(SS_TYPE_INT, sval3);
 			break;
 		// Bitwise operators:
